Give consistent sources to each plugin when migrating
Currently, the list of sources in a project is recalculated after each
plugin is run. Some plugins, like "explicitAny" and
"declareMissingClasProperties", expand the scope of a project by pulling
dependencies of the file they're modifying into the project. This means
that the order in which plugins run impacts the files that ts-migrate
touches.

This commit finds sources for a pronect once, before any plugins are
run, and re-uses them for each plugin. This prevents files from being
pulled in, which is particularly surprising if the '--sources' flag is
in use.